refactor: split native/Sources/main.swift into modules (closes #9)
Split the 1,829-line single-file Swift CLI into logical modules:

- Types.swift: shared types (TargetApp, NativeError, ParsedOptions,
  DescribeOptions, SearchOptions)
- Utils.swift: shared utilities (argument parsing, process execution,
  accessibility helpers, mouse/keyboard events, coordinate conversion,
  text formatting, element search/describe)
- Commands/UICommands.swift: describe-ui, search-ui, tap, type, swipe,
  scroll handlers
- Commands/InputCommands.swift: key, key-sequence, key-combo, button,
  touch, gesture handlers
- Commands/MediaCommands.swift: screenshot, screenshot-app, record-video,
  stream-video handlers
- Commands/SystemCommands.swift: list-apps, list-windows, activate-app,
  menu-action, get-focused-app, clipboard handlers
- Commands/WindowCommands.swift: right-click, drag-drop handlers
- main.swift: entry point with runParsed() dispatch and printHelp()

Binary behavior is identical. No command names, argument parsing, or
output formats were changed. Package.swift auto-discovers all source
files in the Sources/ directory.
